Refactor add new feature
