+++ packages/core/src/index.ts
@@ -10,6 +10,7 @@ export type { ServiceProvider } from './types/ServiceProvider'
 export { DiagramBuilder } from './DiagramBuilder'
 export { ComputerFactory } from './ComputerFactory'
 export { Diagram } from './Diagram'
+export type { Node } from './types/Node'
 export type { Link } from './types/Link'
 export { PositionGuesser } from './PositionGuesser'
 export { LinkGuesser } from './LinkGuesser'
